requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
mlflow==2.16.0
fastapi==0.112.2
uvicorn==0.30.6
pyyaml==6.0.2
joblib==1.4.2
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
streamlit==1.38.0
plotly==5.24.0
requests==2.32.3